requirements.txt
# Core dependencies
feedparser==6.0.10      # RSS feed parsing
requests==2.31.0        # HTTP client for APIs
beautifulsoup4==4.12.2  # HTML parsing
python-dotenv==1.0.0    # Load environment variables from .env
pyyaml==6.0.1           # Parse config YAML files
rapidfuzz==3.6.1        # Fuzzy string matching for article deduplication

# Agent framework (LangGraph + Ollama)
langgraph>=0.2.0        # Agent orchestration with state management
langchain-ollama>=0.2.0 # Ollama LLM integration
langchain-core>=0.3.0   # LangChain core abstractions

# MCP (Model Context Protocol)
mcp>=1.0.0              # MCP client for tool communication

# Development and testing
pytest==8.0.0           # Test framework
pytest-mock==3.12.0     # Mocking utilities for tests
pytest-asyncio>=0.23.0  # Async test support
black==24.1.0           # Code formatting
flake8==7.0.0           # Linting
